XRDDEV-87 Update copyright owner and package naming.
- Update NIIS as the copyright owner.
- Update repository URLs in pom.xml.
- Update package naming from com.pkrete.xrd4j.tools to org.niis.
- Add license header to source code files using Maven License plugin.
