+++ src/src/main/license/LICENSE.txt
@@ -1,6 +1,6 @@
 The MIT License (MIT)
 
-Copyright (c) 2015 Petteri Kivimäki
+Copyright (c) 2018 Nordic Institute for Interoperability Solutions (NIIS)
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
