+++ src/src/main/webapp/WEB-INF/web.xml
@@ -2,7 +2,7 @@
 <web-app version="2.5" xmlns="http://java.sun.com/xml/ns/javaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd">
     <servlet>
         <servlet-name>Endpoint</servlet-name>
-        <servlet-class>com.pkrete.xrd4j.tools.testservice.Endpoint</servlet-class>
+        <servlet-class>org.niis.testservice.Endpoint</servlet-class>
         <load-on-startup>1</load-on-startup>
     </servlet>
     <servlet-mapping>
